refactor: remove coordinator API layer and simplify Redux
Delete coordinator services (walletApi, transactionApi, signatureApi),
old contexts (MidenClientContext), old hooks (useMidenSdk, useWalletData,
useFungibleAssets), unused components (ConnectWalletModal, WalletInfo,
DynamicWalletButton, etc.), and old type files. Simplify Redux store
to only wallet form slice. Remove transactionSlice, signatureSlice,
and walletStatsSlice.
