--- bin/coordinator-frontend/src/contexts/MidenClientContext.tsx
@@ -1,93 +1,0 @@
-'use client';
-
-import React, { createContext, useContext, useState, useEffect, ReactNode } from 'react';
-import { MidenWebClientHandle } from '../../lib/miden-client';
-
-interface MidenClientContextType {
-  handle: MidenWebClientHandle | null;
-  status: string;
-  isRunning: boolean;
-  isInitialized: boolean;
-  error: string | null;
-  reinitialize: () => Promise<void>;
-}
-
-const MidenClientContext = createContext<MidenClientContextType | undefined>(undefined);
-
-export function MidenClientProvider({ children }: { children: ReactNode }) {
-  const [handle, setHandle] = useState<MidenWebClientHandle | null>(null);
-  const [status, setStatus] = useState<string>('');
-  const [isRunning, setIsRunning] = useState(false);
-  const [isInitialized, setIsInitialized] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const initializationRef = React.useRef(false);
-
-  const reinitialize = async () => {
-    if (initializationRef.current || isRunning || isInitialized) {
-      console.info('MidenClient: Initialization skipped (already running or complete)');
-      return;
-    }
-
-    console.info('MidenClient: Starting initialization');
-    initializationRef.current = true;
-    setIsRunning(true);
-    setStatus('Initializing Miden client...');
-    setError(null);
-
-    try {
-      const newHandle = new MidenWebClientHandle();
-      setHandle(newHandle);
-
-      const success = await newHandle.initialize();
-      if (success) {
-        console.info('MidenClient: Initialized successfully');
-        setStatus('Miden client initialized successfully!');
-        setIsInitialized(true);
-      } else {
-        console.error('MidenClient: Initialization failed');
-        setStatus('Failed to initialize Miden client');
-        setIsInitialized(false);
-        setError('Initialization failed');
-        initializationRef.current = false; // Allow retry on failure
-      }
-    } catch (err) {
-      const errorMessage = err instanceof Error ? err.message : 'Unknown error';
-      console.error('MidenClient: Initialization error:', err);
-      setStatus(`Error: ${errorMessage}`);
-      setIsInitialized(false);
-      setError(errorMessage);
-      initializationRef.current = false; // Allow retry on error
-    } finally {
-      setIsRunning(false);
-    }
-  };
-
-  // Initialize once on mount
-  useEffect(() => {
-    reinitialize();
-  }, []);
-
-  return (
-    <MidenClientContext.Provider
-      value={{
-        handle,
-        status,
-        isRunning,
-        isInitialized,
-        error,
-        reinitialize,
-      }}
-    >
-      {children}
-    </MidenClientContext.Provider>
-  );
-}
-
-export function useMidenClient() {
-  const context = useContext(MidenClientContext);
-  if (context === undefined) {
-    throw new Error('useMidenClient must be used within a MidenClientProvider');
-  }
-  return context;
-}
