--- bin/coordinator-frontend/src/hooks/useFungibleAssets.ts
@@ -1,74 +1,0 @@
-import { useState, useEffect, useCallback } from 'react';
-import { useMidenSdk } from './useMidenSdk';
-import { useMidenClient } from '../contexts/MidenClientContext';
-import { AccountId, NetworkId, AccountInterface } from '@demox-labs/miden-sdk';
-import { FungibleAsset } from '@/types';
-
-export const useFungibleAssets = () => {
-  const { Miden } = useMidenSdk();
-  const { handle, isInitialized } = useMidenClient();
-  const [fungibleAssets, setFungibleAssets] = useState<FungibleAsset[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  const getFungibleAssets = useCallback(async () => {
-    try {
-      setIsLoading(true);
-      setError(null);
-
-      const currentWalletId = localStorage.getItem("currentWalletId");
-      if (!currentWalletId || !Miden || !handle || !isInitialized) {
-        setIsLoading(false);
-        return;
-      }
-
-      const webClient = handle.getWebClient();
-      if (!webClient) {
-        setIsLoading(false);
-        return;
-      }
-
-      const address = Miden.Address.fromBech32(currentWalletId);
-      const accountId = address.accountId();
-
-      // Sync state before getting/importing account
-      await webClient.syncState();
-
-      let account = await webClient.getAccount(accountId);
-      if (!account) {
-        await webClient.importAccountById(accountId);
-        account = await webClient.getAccount(accountId);
-      }
-
-      if (account) {
-        const assetVault = account.vault();
-        const assets = assetVault.fungibleAssets();
-
-        const assetsWithBalance = assets.map(asset => {
-          const faucetId = asset.faucetId().toBech32(NetworkId.Testnet, AccountInterface.BasicWallet);
-          const balance = assetVault.getBalance(asset.faucetId()).toString();
-          return { faucetId, balance };
-        });
-
-        setFungibleAssets(assetsWithBalance);
-      }
-    } catch (error) {
-      console.error('Error getting fungible assets:', error);
-      setError(error instanceof Error ? error.message : 'Failed to fetch fungible assets');
-    } finally {
-      setIsLoading(false);
-    }
-  }, [Miden, handle, isInitialized]);
-
-  // Load fungible assets on component mount
-  useEffect(() => {
-    getFungibleAssets();
-  }, [getFungibleAssets]);
-
-  return {
-    fungibleAssets,
-    isLoading,
-    error,
-    getFungibleAssets
-  };
-};
