--- bin/coordinator-frontend/src/services/index.ts
@@ -1,4 +1,0 @@
-// Export all API functions from a single location
-export * from './walletApi';
-export * from './transactionApi';
-export * from './signatureApi';
