+++ bin/coordinator-frontend/src/store/slices/walletSlice.ts
@@ -1,17 +1,19 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { WalletFormData, Approver, MultisigAccount } from '../../types';
+
+export interface WalletFormData {
+  walletName: string;
+  signatureThreshold: string;
+  totalSigners: string;
+  network: string;
+  signerAddresses: string[];
+  signerPublicKeys: string[];
+}
 
 interface WalletState {
   formData: WalletFormData;
   currentStep: number;
   isLoading: boolean;
   error: string | null;
-  approvers: Approver[];
-  approversLoading: boolean;
-  approversError: string | null;
-  walletData: MultisigAccount | null;
-  walletDataLoading: boolean;
-  walletDataError: string | null;
 }
 
 const initialState: WalletState = {
@@ -26,12 +28,6 @@ const initialState: WalletState = {
   currentStep: 1,
   isLoading: false,
   error: null,
-  approvers: [],
-  approversLoading: false,
-  approversError: null,
-  walletData: null,
-  walletDataLoading: false,
-  walletDataError: null,
 };
 
 const walletSlice = createSlice({
@@ -43,8 +39,7 @@ const walletSlice = createSlice({
     },
     updateFormField: (state, action: PayloadAction<{ field: keyof WalletFormData; value: string }>) => {
       const { field, value } = action.payload;
-      if (field === 'signerAddresses') {
-        // Handle signer addresses array separately
+      if (field === 'signerAddresses' || field === 'signerPublicKeys') {
         return;
       }
       (state.formData[field] as string) = value;
@@ -70,7 +65,6 @@ const walletSlice = createSlice({
       state.currentStep = action.payload;
     },
     resetForm: (state) => {
-   
       state.formData = initialState.formData;
       state.currentStep = 1;
       state.error = null;
@@ -80,38 +74,6 @@ const walletSlice = createSlice({
     },
     setError: (state, action: PayloadAction<string | null>) => {
       state.error = action.payload;
-    },
-    setApprovers: (state, action: PayloadAction<Approver[]>) => {
-      state.approvers = action.payload;
-      state.approversError = null;
-    },
-    setApproversLoading: (state, action: PayloadAction<boolean>) => {
-      state.approversLoading = action.payload;
-    },
-    setApproversError: (state, action: PayloadAction<string | null>) => {
-      state.approversError = action.payload;
-      state.approversLoading = false;
-    },
-    clearApprovers: (state) => {
-      state.approvers = [];
-      state.approversError = null;
-      state.approversLoading = false;
-    },
-    setWalletData: (state, action: PayloadAction<MultisigAccount>) => {
-      state.walletData = action.payload;
-      state.walletDataError = null;
-    },
-    setWalletDataLoading: (state, action: PayloadAction<boolean>) => {
-      state.walletDataLoading = action.payload;
-    },
-    setWalletDataError: (state, action: PayloadAction<string | null>) => {
-      state.walletDataError = action.payload;
-      state.walletDataLoading = false;
-    },
-    clearWalletData: (state) => {
-      state.walletData = null;
-      state.walletDataError = null;
-      state.walletDataLoading = false;
     },
   },
 });
@@ -127,15 +89,7 @@ export const {
   resetForm,
   setLoading,
   setError,
-  setApprovers,
-  setApproversLoading,
-  setApproversError,
-  clearApprovers,
-  setWalletData,
-  setWalletDataLoading,
-  setWalletDataError,
-  clearWalletData,
 } = walletSlice.actions;
 
-export type { WalletState, WalletFormData };
-export default walletSlice.reducer; 
+export type { WalletState };
+export default walletSlice.reducer;
