+++ bin/coordinator-frontend/src/types/index.ts
@@ -1,8 +1,4 @@
 // Export all types from a single location
-export * from './wallet';
-export * from './transaction';
-export * from './signature';
 export * from './components';
-export * from './hooks';
 export * from './store';
-export * from './notes'; 
+export * from './psm';
