--- bin/coordinator-frontend/src/types/signature.ts
@@ -1,6 +1,0 @@
-// Signature-related types
-export interface AddSignatureRequest {
-  tx_id: string;
-  approver: string;
-  signature: string;
-}
